Some fixes.
git-svn-id: https://ilk.uvt.nl/svn/trunk/sources/wopr@17351 12f355fe-0486-481a-ad91-c297ab22b4e3

+++ include/wopr/PDT.h
@@ -137,7 +137,7 @@ class History {
   }
 
   void dump() {
-    for ( int i = 0; i < his.size(); i++ ) {
+    for ( size_t i = 0; i < his.size(); i++ ) {
       std::cerr << the_time() << ": (dump)(" << i << ")=" << his.at(i)->toString() << std::endl;
     }
   }
